@@ -5,9 +5,10 @@
 using System;
 using System.Collections.Immutable;
 using System.ComponentModel.DataAnnotations;
+using System.IO;
 using System.Net.Http;
-using System.Net.Http.Formatting;
 using System.Net.Http.Headers;
+using System.Text;
 using System.Threading;
 using System.Threading.Tasks;
 using DotNet.Status.Web.Options;
@@ -28,14 +29,14 @@ namespace DotNet.Status.Web.Controllers
     [Route("api/deployment")]
     public class DeploymentController : ControllerBase
     {
-        private readonly IHostingEnvironment _env;
+        private readonly IHostEnvironment _env;
         private readonly IOptionsMonitor<GrafanaOptions> _grafanaOptions;
         private readonly ILogger<DeploymentController> _logger;
 
         public DeploymentController(
             IOptionsMonitor<GrafanaOptions> grafanaOptions,
             ILogger<DeploymentController> logger,
-            IHostingEnvironment env)
+            IHostEnvironment env)
         {
             _grafanaOptions = grafanaOptions;
             _logger = logger;
@@ -66,14 +67,13 @@ namespace DotNet.Status.Web.Controllers
                             request.Headers.Accept.Add(new MediaTypeWithQualityHeaderValue("application/json"));
                             request.Headers.Authorization =
                                 new AuthenticationHeaderValue("Bearer", grafanaOptions.ApiToken);
-                            request.Content =
-                                new ObjectContent<NewGrafanaAnnotationRequest>(content, s_grafanaFormatter);
+                            request.Content = CreateObjectContent(content);
 
                             using (HttpResponseMessage response = await client.SendAsync(request, CancellationToken.None))
                             {
                                 _logger.LogTrace("Response from grafana {responseCode} {reason}", response.StatusCode, response.ReasonPhrase);
                                 response.EnsureSuccessStatusCode();
-                                return await response.Content.ReadAsAsync<NewGrafanaAnnotationResponse>(s_grafanaFormatters, CancellationToken.None);
+                                return await ReadObjectContent<NewGrafanaAnnotationResponse>(response.Content);
                             }
                         }
                     },
@@ -130,8 +130,7 @@ namespace DotNet.Status.Web.Controllers
                             request.Headers.Accept.Add(new MediaTypeWithQualityHeaderValue("application/json"));
                             request.Headers.Authorization =
                                 new AuthenticationHeaderValue("Bearer", grafanaOptions.ApiToken);
-                            request.Content =
-                                new ObjectContent<NewGrafanaAnnotationRequest>(content, s_grafanaFormatter);
+                            request.Content = CreateObjectContent(content);
                             using (HttpResponseMessage response = await client.SendAsync(request, CancellationToken.None))
                             {
                                 _logger.LogTrace("Response from grafana {responseCode} {reason}", response.StatusCode, response.ReasonPhrase);
@@ -162,18 +161,28 @@ namespace DotNet.Status.Web.Controllers
             }
             return table;
         }
-        
-        private static readonly MediaTypeFormatter s_grafanaFormatter = new JsonMediaTypeFormatter
-        {
-            SerializerSettings =
-            {
-                ContractResolver = new CamelCasePropertyNamesContractResolver(),
-                NullValueHandling = NullValueHandling.Ignore,
-            }
-        };
 
-        private static readonly ImmutableArray<MediaTypeFormatter> s_grafanaFormatters =
-            ImmutableArray.Create(s_grafanaFormatter);
+        private static StringContent CreateObjectContent(object content)
+        {
+            StringWriter writer = new StringWriter();
+            s_grafanaSerializer.Serialize(writer, content);
+            return new StringContent(writer.ToString(), Encoding.UTF8, "application/json");
+        }
+
+        private static async Task<T> ReadObjectContent<T>(HttpContent content)
+        {
+            using (var streamReader = new StreamReader(await content.ReadAsStreamAsync()))
+            using (var jsonReader = new JsonTextReader(streamReader))
+            {
+                return s_grafanaSerializer.Deserialize<T>(jsonReader);
+            }
+        }
+
+        private static readonly JsonSerializer s_grafanaSerializer = new JsonSerializer
+        {
+            ContractResolver = new CamelCasePropertyNamesContractResolver(),
+            Formatting = Formatting.None,
+        };
     }
 
     public class DeploymentStartRequest
