@@ -9,6 +9,7 @@ using Microsoft.ApplicationInsights.AspNetCore.Extensions;
 using Microsoft.AspNetCore.Hosting;
 using Microsoft.AspNetCore.Html;
 using Microsoft.AspNetCore.Mvc.RazorPages;
+using Microsoft.Extensions.Hosting;
 using Microsoft.Extensions.Options;
 using Newtonsoft.Json.Linq;
 
@@ -16,9 +17,9 @@ namespace Maestro.Web.Pages
 {
     public class IndexModel : PageModel
     {
-        public IHostingEnvironment Environment { get; }
+        public IWebHostEnvironment Environment { get; }
 
-        public IndexModel(IHostingEnvironment environment, IOptions<ApplicationInsightsServiceOptions> applicationInsightsOptions)
+        public IndexModel(IWebHostEnvironment environment, IOptions<ApplicationInsightsServiceOptions> applicationInsightsOptions)
         {
             Environment = environment;
             InstrumentationKey = applicationInsightsOptions.Value.InstrumentationKey;
