@@ -4,6 +4,7 @@
 
 using System;
 using System.Collections.Generic;
+using System.Globalization;
 using System.Linq;
 using System.Net;
 using System.Net.Http;
@@ -24,10 +25,10 @@ using Microsoft.AspNetCore.ApiVersioning;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.DataProtection;
+using Microsoft.AspNetCore.Hosting;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Rewrite;
-using Microsoft.AspNetCore.Rewrite.Internal;
 using Microsoft.Azure.KeyVault;
 using Microsoft.Azure.KeyVault.Models;
 using Microsoft.DotNet.DarcLib;
@@ -43,10 +44,13 @@ using Newtonsoft.Json;
 using Microsoft.DotNet.GitHub.Authentication;
 using Microsoft.DotNet.Kusto;
 using Microsoft.Azure.Services.AppAuthentication;
+using Microsoft.OpenApi.Models;
+using Newtonsoft.Json.Converters;
+using Newtonsoft.Json.Serialization;
 
 namespace Maestro.Web
 {
-    public partial class Startup
+    public partial class Startup : StartupBase
     {
         // https://github.com/dotnet/core-eng/issues/6819
         // TODO: Remove once the repo in this list is ready to onboard to yaml publishing.
@@ -90,7 +94,7 @@ namespace Maestro.Web
             };
         }
 
-        public Startup(IConfiguration configuration, IHostingEnvironment env)
+        public Startup(IConfiguration configuration, IHostEnvironment env)
         {
             HostingEnvironment = env;
             Configuration = configuration;
@@ -98,11 +102,10 @@ namespace Maestro.Web
 
         public static readonly TimeSpan LoginCookieLifetime = new TimeSpan(days: 120, hours: 0, minutes: 0, seconds: 0);
 
-        public IHostingEnvironment HostingEnvironment { get; set; }
+        public IHostEnvironment HostingEnvironment { get; }
         public IConfiguration Configuration { get; }
 
-        // This method gets called by the runtime. Use this method to add services to the container.
-        public void ConfigureServices(IServiceCollection services)
+        public override void ConfigureServices(IServiceCollection services)
         {
             if (HostingEnvironment.IsDevelopment())
             {
@@ -153,17 +156,15 @@ namespace Maestro.Web
                     options.UseSqlServer(Configuration.GetSection("BuildAssetRegistry")["ConnectionString"]);
                 });
 
-            services.AddMvc()
-                .SetCompatibilityVersion(CompatibilityVersion.Version_2_1)
+            services.AddRazorPages(options =>
+                {
+                    options.Conventions.AuthorizeFolder("/", MsftAuthorizationPolicyName);
+                    options.Conventions.AllowAnonymousToPage("/Index");
+                    options.Conventions.AllowAnonymousToPage("/Error");
+                    options.Conventions.AllowAnonymousToPage("/SwaggerUi");
+                })
+                .SetCompatibilityVersion(CompatibilityVersion.Version_3_0)
                 .AddFluentValidation(options => options.RegisterValidatorsFromAssemblyContaining<Startup>())
-                .AddRazorPagesOptions(
-                    options =>
-                    {
-                        options.Conventions.AuthorizeFolder("/", MsftAuthorizationPolicyName);
-                        options.Conventions.AllowAnonymousToPage("/Index");
-                        options.Conventions.AllowAnonymousToPage("/Error");
-                        options.Conventions.AllowAnonymousToPage("/SwaggerUi");
-                    })
                 .AddGitHubWebHooks()
                 .AddApiPagination()
                 .AddCookieTempDataProvider(
@@ -174,7 +175,21 @@ namespace Maestro.Web
                         options.Cookie.IsEssential = true;
                     });
 
-            services.AddSingleton<IConfiguration>(Configuration);
+            services.AddControllers()
+                .AddNewtonsoftJson(options =>
+                {
+                    options.SerializerSettings.ContractResolver = new CamelCasePropertyNamesContractResolver();
+                    options.SerializerSettings.Converters.Add(new StringEnumConverter
+                        {NamingStrategy = new CamelCaseNamingStrategy()});
+                    options.SerializerSettings.Converters.Add(
+                        new IsoDateTimeConverter
+                        {
+                            DateTimeFormat = "yyyy-MM-ddTHH:mm:ssZ",
+                            DateTimeStyles = DateTimeStyles.AdjustToUniversal
+                        });
+                });
+
+            services.AddSingleton(Configuration);
 
             ConfigureAuthServices(services);
 
@@ -220,7 +235,7 @@ namespace Maestro.Web
             // in such a way that will work with sizing.
             services.AddSingleton<DarcRemoteMemoryCache>();
 
-            services.AddSingleton<IRemoteFactory, DarcRemoteFactory>();
+            services.AddScoped<IRemoteFactory, DarcRemoteFactory>();
             services.AddSingleton(typeof(IActorProxyFactory<>), typeof(ActorProxyFactory<>));
 
             services.AddMergePolicies();
@@ -233,8 +248,8 @@ namespace Maestro.Web
                 async ctx =>
                 {
                     var result = new ApiError("An error occured.");
-                    MvcJsonOptions jsonOptions =
-                        ctx.RequestServices.GetRequiredService<IOptions<MvcJsonOptions>>().Value;
+                    MvcNewtonsoftJsonOptions jsonOptions =
+                        ctx.RequestServices.GetRequiredService<IOptions<MvcNewtonsoftJsonOptions>>().Value;
                     string output = JsonConvert.SerializeObject(result, jsonOptions.SerializerSettings);
                     ctx.Response.StatusCode = (int)HttpStatusCode.InternalServerError;
                     await ctx.Response.WriteAsync(output, Encoding.UTF8);
@@ -277,21 +292,51 @@ namespace Maestro.Web
         {
             app.UseExceptionHandler(ConfigureApiExceptions);
 
-            app.UseAuthentication();
-
-            if (HostingEnvironment.IsDevelopment() && !Program.RunningInServiceFabric())
+            if (HostingEnvironment.IsDevelopment() &&
+                !Program.RunningInServiceFabric() &&
+                !string.Equals(
+                    Configuration["ForceLocalApi"],
+                    true.ToString(),
+                    StringComparison.OrdinalIgnoreCase))
             {
                 // Redirect api requests to prod when running locally outside of service fabric
                 // This is for the `ng serve` local debugging case for the website
                 app.MapWhen(
-                    ctx => IsGet(ctx) && ctx.Request.Path.StartsWithSegments("/api") && ctx.Request.Path != "/api/swagger.json",
-                    a =>
-                    {
-                        a.Run(ApiRedirectHandler);
-                    });
+                    ctx => IsGet(ctx) &&
+                        ctx.Request.Path.StartsWithSegments("/api") &&
+                        ctx.Request.Path != "/api/swagger.json",
+                    a => { a.Run(ApiRedirectHandler); });
             }
 
-            app.UseMvc();
+            app.UseSwagger(
+                options =>
+                {
+                    options.SerializeAsV2 = true;
+                    options.RouteTemplate = "api/{documentName}/swagger.json";
+                    options.PreSerializeFilters.Add(
+                        (doc, req) =>
+                        {
+                            bool http = HostingEnvironment.IsDevelopment() && !Program.RunningInServiceFabric();
+                            doc.Servers = new List<OpenApiServer>
+                            {
+                                new OpenApiServer
+                                {
+                                    Url = $"{(http ? "http" : "https")}://{req.Host.Value}/",
+                                },
+                            };
+
+                            req.HttpContext.Response.Headers["Access-Control-Allow-Origin"] = "*";
+                        });
+                });
+            
+            app.UseRouting();
+            app.UseAuthentication();
+            app.UseAuthorization();
+            app.UseEndpoints(e =>
+            {
+                e.MapRazorPages();
+                e.MapControllers();
+            });
 
             app.Use(
                 (ctx, next) =>
@@ -305,57 +350,41 @@ namespace Maestro.Web
 
                     return next();
                 });
-
-            app.UseSwagger(
-                options =>
-                {
-                    options.RouteTemplate = "api/{documentName}/swagger.json";
-                    options.PreSerializeFilters.Add(
-                        (doc, req) =>
-                        {
-                            doc.Host = req.Host.Value;
-                            if (HostingEnvironment.IsDevelopment() && !Program.RunningInServiceFabric())
-                            {
-                                doc.Schemes = new List<string> { "http" };
-                            }
-                            else
-                            {
-                                doc.Schemes = new List<string> { "https" };
-                            }
-
-                            req.HttpContext.Response.Headers["Access-Control-Allow-Origin"] = "*";
-                        });
-                });
         }
 
         // The whole api, only allowing GET requests, with all urls prefixed with _
         private void ConfigureCookieAuthedApi(IApplicationBuilder app)
         {
             app.UseExceptionHandler(ConfigureApiExceptions);
-            app.UseAuthentication();
 
-            app.UseRewriter(new RewriteOptions
-            {
-                Rules =
-                {
-                    new RewriteRule("^_/(.*)", "$1", true),
-                },
-            });
+            app.UseRouting();
+            app.UseAuthentication();
+            app.UseAuthorization();
+
+            app.UseRewriter(new RewriteOptions().AddRewrite("^_/(.*)", "$1", true));
 
             // Redirect the entire cookie-authed api if it is in settings.
             if (DoApiRedirect)
             {
                 // when told to not redirect by the request, don't do it.
-                app.MapWhen(ctx => ctx.Request.Cookies.TryGetValue("Skip-Api-Redirect", out _), a =>
-                {
-                    a.UseMvc();
-                });
+                app.MapWhen(ctx => ctx.Request.Cookies.TryGetValue("Skip-Api-Redirect", out _),
+                    a =>
+                    {
+                        a.UseRouting();
+                        a.UseEndpoints(e =>
+                        {
+                            e.MapControllers();
+                        });
+                    });
 
                 app.Run(ApiRedirectHandler);
             }
             else
             {
-                app.UseMvc();
+                app.UseEndpoints(e =>
+                {
+                    e.MapControllers();
+                });
             }
 
         }
@@ -365,9 +394,9 @@ namespace Maestro.Web
             return string.Equals(context.Request.Method, "get", StringComparison.OrdinalIgnoreCase);
         }
 
-        public void Configure(IApplicationBuilder app, IHostingEnvironment env)
+        public override void Configure(IApplicationBuilder app)
         {
-            if (env.IsDevelopment())
+            if (HostingEnvironment.IsDevelopment())
             {
                 app.UseDeveloperExceptionPage();
             }
@@ -409,16 +438,10 @@ namespace Maestro.Web
                     return next();
                 });
 
-            if (env.IsDevelopment() && !Program.RunningInServiceFabric())
+            if (HostingEnvironment.IsDevelopment() && !Program.RunningInServiceFabric())
             {
                 // In local dev with the `ng serve` scenario, just redirect /_/api to /api
-                app.UseRewriter(new RewriteOptions
-                {
-                    Rules =
-                    {
-                        new RewriteRule("^_/(.*)", "$1", true),
-                    },
-                });
+                app.UseRewriter(new RewriteOptions().AddRewrite("^_/(.*)", "$1", true));
             }
 
             app.MapWhen(ctx => ctx.Request.Path.StartsWithSegments("/api"), ConfigureApi);
@@ -433,22 +456,25 @@ namespace Maestro.Web
             app.UseStatusCodePagesWithReExecute("/Error", "?code={0}");
             app.UseCookiePolicy();
             app.UseStaticFiles();
-            app.UseAuthentication();
 
-            app.UseMvc();
+            app.UseRouting();
+            app.UseAuthentication();
+            app.UseAuthorization();
+
+            app.UseEndpoints(e =>
+                {
+                    e.MapRazorPages();
+                    e.MapControllers();
+                }
+            );
             app.MapWhen(IsGet, AngularIndexHtmlRedirect);
         }
 
         private static void AngularIndexHtmlRedirect(IApplicationBuilder app)
         {
-            app.UseRewriter(new RewriteOptions
-            {
-                Rules =
-                {
-                    new RewriteRule(".*", "Index", true),
-                },
-            });
-            app.UseMvc();
+            app.UseRewriter(new RewriteOptions().AddRewrite(".*", "Index", true));
+            app.UseRouting();
+            app.UseEndpoints(e => { e.MapRazorPages(); });
         }
     }
 }
