@@ -5,16 +5,15 @@
 using Maestro.AzureDevOps;
 using Maestro.Data;
 using Maestro.Data.Models;
-using Microsoft.AspNetCore.Hosting;
 using Microsoft.DotNet.DarcLib;
 using Microsoft.EntityFrameworkCore;
-using Microsoft.EntityFrameworkCore.Internal;
 using Microsoft.Extensions.DependencyInjection;
 using Moq;
 using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
+using Microsoft.Extensions.Hosting;
 using Xunit;
 
 namespace FeedCleanerService.Tests
@@ -22,7 +21,7 @@ namespace FeedCleanerService.Tests
     public class FeedCleanerServiceTests : IDisposable
     {
         private readonly Lazy<BuildAssetRegistryContext> _context;
-        private readonly Mock<IHostingEnvironment> Env;
+        private readonly Mock<IHostEnvironment> Env;
         private readonly ServiceProvider Provider;
         private readonly IServiceScope Scope;
         private readonly Dictionary<string, AzureDevOpsFeed> Feeds;
@@ -37,11 +36,15 @@ namespace FeedCleanerService.Tests
         public FeedCleanerServiceTests()
         {
             var services = new ServiceCollection();
-            Env = new Mock<IHostingEnvironment>(MockBehavior.Strict);
+            Env = new Mock<IHostEnvironment>(MockBehavior.Strict);
             services.AddSingleton(Env.Object);
             services.AddLogging();
             services.AddDbContext<BuildAssetRegistryContext>(
-                options => { options.UseInMemoryDatabase("BuildAssetRegistry"); });
+                options =>
+                {
+                    options.UseInMemoryDatabase("BuildAssetRegistry");
+                    options.EnableServiceProviderCaching(false);
+                });
             services.Configure<FeedCleanerOptions>(
                 (options) =>
                 {
@@ -104,7 +107,7 @@ namespace FeedCleanerService.Tests
             await cleaner.CleanManagedFeedsAsync();
 
             // Check the assets for the feed where all packages were released.
-            var assetsInDeletedFeed = GetContext().Assets.Where(a => a.Locations.Any(l => l.Location.Contains(FeedWithAllPackagesReleasedName))).ToList();
+            var assetsInDeletedFeed = _context.Value.Assets.Where(a => a.Locations.Any(l => l.Location.Contains(FeedWithAllPackagesReleasedName))).ToList();
             Assert.Equal(4, assetsInDeletedFeed.Count);
             Assert.Contains(assetsInDeletedFeed,
                 a => a.Name.Equals("Newtonsoft.Json") &&
@@ -117,7 +120,7 @@ namespace FeedCleanerService.Tests
                 !a.Locations.Any(l => l.Location.Contains(ReleaseFeedName)));
 
             // "releasedPackage1" should've been released and have its location updated to the released feed.
-            var assetsInRemainingFeed = GetContext().Assets.Where(a => a.Locations.Any(l => l.Location.Contains(FeedWithUnreleasedPackagesName))).ToList();
+            var assetsInRemainingFeed = _context.Value.Assets.Where(a => a.Locations.Any(l => l.Location.Contains(FeedWithUnreleasedPackagesName))).ToList();
             Assert.Equal(2, assetsInRemainingFeed.Count);
             var releasedAssets = assetsInRemainingFeed.Where(a => a.Locations.Any(l => l.Location.Contains(ReleaseFeedName))).ToList();
             Assert.Single(releasedAssets);
@@ -208,7 +211,7 @@ namespace FeedCleanerService.Tests
                 }
             }
 
-            var context = GetContext();
+            var context = _context.Value;
             context.Assets.AddRange(assets);
             context.SaveChanges();
         }
