@@ -429,10 +429,10 @@ namespace Maestro.Web.Api.v2018_07_16.Controllers
             // - Target branch
             // - Not the same subscription id (for updates)
             return await _context.Subscriptions.FirstOrDefaultAsync(sub =>
-                sub.SourceRepository.Equals(updatedOrNewSubscription.SourceRepository, StringComparison.OrdinalIgnoreCase) &&
+                sub.SourceRepository == updatedOrNewSubscription.SourceRepository &&
                 sub.ChannelId == updatedOrNewSubscription.Channel.Id &&
-                sub.TargetRepository.Equals(updatedOrNewSubscription.TargetRepository, StringComparison.OrdinalIgnoreCase) &&
-                sub.TargetBranch.Equals(updatedOrNewSubscription.TargetBranch, StringComparison.OrdinalIgnoreCase) &&
+                sub.TargetRepository == updatedOrNewSubscription.TargetRepository &&
+                sub.TargetBranch == updatedOrNewSubscription.TargetBranch &&
                 sub.Id != updatedOrNewSubscription.Id);
         }
     }
