@@ -4,21 +4,19 @@
 
 using System;
 using System.Collections.Generic;
-using System.Globalization;
 using System.IO;
 using System.Linq;
 using System.Reflection;
 using Microsoft.AspNetCore.ApiPagination;
 using Microsoft.AspNetCore.ApiVersioning;
 using Microsoft.AspNetCore.ApiVersioning.Swashbuckle;
-using Microsoft.AspNetCore.Hosting;
-using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
-using Newtonsoft.Json.Converters;
+using Microsoft.OpenApi;
+using Microsoft.OpenApi.Interfaces;
+using Microsoft.OpenApi.Models;
+using Microsoft.OpenApi.Writers;
 using Newtonsoft.Json.Linq;
-using Newtonsoft.Json.Serialization;
-using Swashbuckle.AspNetCore.Swagger;
 
 namespace Maestro.Web
 {
@@ -32,25 +30,14 @@ namespace Maestro.Web
                 {
                     info.Description =
                         "The Web API enabling access to the Maestro++ service that supports the [.NET Core Dependency Flow infrastructure](https://github.com/dotnet/arcade/blob/master/Documentation/DependenciesFlowPlan.md).";
-                    info.Contact = new Contact
+                    info.Contact = new OpenApiContact
                     {
                         Name = ".NET Core Engineering",
                         Email = "dnceng@microsoft.com",
                     };
                 });
-            services.Configure<MvcJsonOptions>(
-                options =>
-                {
-                    options.SerializerSettings.ContractResolver = new CamelCasePropertyNamesContractResolver();
-                    options.SerializerSettings.Converters.Add(new StringEnumConverter {CamelCaseText = true});
-                    options.SerializerSettings.Converters.Add(
-                        new IsoDateTimeConverter
-                        {
-                            DateTimeFormat = "yyyy-MM-ddTHH:mm:ssZ",
-                            DateTimeStyles = DateTimeStyles.AdjustToUniversal
-                        });
-                });
 
+            services.AddSwaggerGenNewtonsoftSupport();
             services.AddSwaggerGen(
                 options =>
                 {
@@ -67,16 +54,21 @@ namespace Maestro.Web
                     options.FilterOperations(
                         (op, ctx) =>
                         {
-                            op.Responses["default"] = new Response
+                            var errorSchema = ctx.SchemaGenerator.GenerateSchema(typeof(ApiError), ctx.SchemaRepository);
+                            op.Responses["default"] = new OpenApiResponse
                             {
                                 Description = "Error",
-                                Schema = ctx.SchemaRegistry.GetOrRegister(typeof(ApiError))
+                                Content = new Dictionary<string, OpenApiMediaType>
+                                {
+                                    ["application/json"] = new OpenApiMediaType
+                                    {
+                                        Schema = errorSchema,
+                                    },
+                                },
                             };
                             // Replace the large list generated by WepApi with just application/json 
                             // We can accept more than just application/json, but the swagger spec defines what we prefer
-                            op.Consumes = appJson;
-                            op.Produces = appJson;
-                            op.OperationId = $"{op.Tags.First()}_{op.OperationId}";
+                            op.OperationId = $"{op.Tags.First().Name}_{op.OperationId}";
                         });
 
                     options.FilterOperations(
@@ -86,10 +78,10 @@ namespace Maestro.Web
                             if (paginated != null)
                             {
                                 // Add an extension that tells the client generator that this operation is paged with first,prev,next,last urls in the Link header.
-                                op.Extensions["x-ms-paginated"] = new
+                                op.Extensions["x-ms-paginated"] = new PaginatedExtension
                                 {
-                                    page = paginated.PageParameterName,
-                                    pageSize = paginated.PageSizeParameterName
+                                    PageParameterName = paginated.PageParameterName,
+                                    PageSizeParameterName = paginated.PageSizeParameterName
                                 };
                             }
                         });
@@ -98,43 +90,53 @@ namespace Maestro.Web
                         (schema, ctx) =>
                         {
                             // Types that are not-nullable in C# should be required
-                            if (ctx.JsonContract is JsonObjectContract contract)
+                            if (schema.Type == "object")
                             {
                                 var required = schema.Required == null
                                     ? new HashSet<string>()
-                                    : new HashSet<string>(schema.Required);
-                                foreach (var property in contract.Properties)
+                                    : new HashSet<string>(schema.Required.Select(ToCamelCase));
+                                schema.Properties =
+                                    schema.Properties.ToDictionary(
+                                        p => ToCamelCase(p.Key),
+                                        p => p.Value);
+                                foreach (var property in schema.Properties.Keys)
                                 {
-                                    var propertyType = property.PropertyType;
-                                    var shouldBeRequired =
-                                        propertyType.IsValueType &&
-                                        !(propertyType.IsConstructedGenericType &&
-                                          propertyType.GetGenericTypeDefinition() == typeof(Nullable<>));
-                                    if (shouldBeRequired)
+                                    var propertyInfo = ctx.Type.GetRuntimeProperties().FirstOrDefault(p =>
+                                        string.Equals(p.Name, property, StringComparison.OrdinalIgnoreCase));
+                                    if (propertyInfo != null)
                                     {
-                                        required.Add(property.PropertyName);
+                                        var propertyType = propertyInfo.PropertyType;
+                                        var shouldBeRequired =
+                                            propertyType.IsValueType &&
+                                            !(propertyType.IsConstructedGenericType &&
+                                              propertyType.GetGenericTypeDefinition() == typeof(Nullable<>));
+                                        if (shouldBeRequired)
+                                        {
+                                            required.Add(property);
+                                        }
                                     }
                                 }
-                                schema.Required = required.ToList();
+
+                                schema.Required = required;
                             }
                         });
 
                     options.MapType<TimeSpan>(
-                        () => new Schema
+                        () => new OpenApiSchema
                         {
                             Type = "string",
                             Format = "duration"
                         });
                     options.MapType<TimeSpan?>(
-                        () => new Schema
+                        () => new OpenApiSchema
                         {
                             Type = "string",
                             Format = "duration"
                         });
-                    options.MapType<JToken>(() => new Schema());
+                    options.MapType<JToken>(() => new OpenApiSchema());
 
-                    options.DescribeAllEnumsAsStrings();
-                    options.SchemaRegistryOptions.UseReferencedDefinitionsForEnums = true;
+                    options.SchemaGeneratorOptions.UseInlineDefinitionsForEnums = false;
+                    options.DescribeAllParametersInCamelCase();
 
                     string xmlPath;
                     if (HostingEnvironment.IsDevelopment())
@@ -154,25 +156,31 @@ namespace Maestro.Web
 
                     options.AddSecurityDefinition(
                         "Bearer",
-                        new ApiKeyScheme
+                        new OpenApiSecurityScheme
                         {
-                            Type = "apiKey",
-                            In = "header",
-                            Name = "Authorization"
-                        });
-
-                    options.FilterDocument(
-                        (doc, ctx) =>
-                        {
-                            doc.Security = new List<IDictionary<string, IEnumerable<string>>>
-                            {
-                                new Dictionary<string, IEnumerable<string>>
-                                {
-                                    ["Bearer"] = Enumerable.Empty<string>()
-                                }
-                            };
+                            Type = SecuritySchemeType.Http,
+                            Scheme = "bearer",
                         });
                 });
+        }
+
+        private static string ToCamelCase(string value)
+        {
+            return value.Substring(0, 1).ToLowerInvariant() + value.Substring(1);
+        }
+    }
+
+    internal class PaginatedExtension : IOpenApiExtension
+    {
+        public string PageParameterName { get; set; }
+        public string PageSizeParameterName { get; set; }
+
+        public void Write(IOpenApiWriter writer, OpenApiSpecVersion specVersion)
+        {
+            writer.WriteStartObject();
+            writer.WriteProperty("page", PageParameterName);
+            writer.WriteProperty("pageSize", PageSizeParameterName);
+            writer.WriteEndObject();
         }
     }
 }
