@@ -2,7 +2,7 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
-using Microsoft.AspNetCore.Mvc.Internal;
+using Microsoft.AspNetCore.Mvc.Razor;
 using Microsoft.AspNetCore.Mvc.Rendering;
 using Microsoft.AspNetCore.Routing;
 
@@ -12,8 +12,8 @@ namespace Maestro.Web.TagHelpers
     {
         public static bool IsCurrentPage(this ViewContext context, string pageName)
         {
-            string normalizedRouteValue = NormalizedRouteValue.GetNormalizedRouteValue(context, "page");
-            string expectedPageValue = ViewEnginePath.CombinePath(normalizedRouteValue, pageName);
+            string normalizedRouteValue = RazorViewEngine.GetNormalizedRouteValue(context, "page");
+            string expectedPageValue = normalizedRouteValue + "/" + pageName;
             RouteData routeData = context.HttpContext.GetRouteData();
             var page = (string) routeData.Values["page"];
             if (page == expectedPageValue)
