@@ -5,6 +5,8 @@
 using System.Collections.Generic;
 using JetBrains.Annotations;
 using Microsoft.AspNetCore.ApiVersioning.Schemes;
+using Microsoft.OpenApi.Any;
+using Microsoft.OpenApi.Models;
 using Swashbuckle.AspNetCore.Swagger;
 using Swashbuckle.AspNetCore.SwaggerGen;
 
@@ -20,23 +22,29 @@ namespace Microsoft.AspNetCore.ApiVersioning.Swashbuckle
             _scheme = scheme;
         }
 
-        public void Apply(Operation operation, OperationFilterContext context)
+        public void Apply(OpenApiOperation operation, OperationFilterContext context)
         {
             string version = context.ApiDescription.ActionDescriptor.RouteValues["version"];
             if (operation.Parameters == null)
             {
-                operation.Parameters = new List<IParameter>();
+                operation.Parameters = new List<OpenApiParameter>();
             }
 
             operation.Parameters.Add(
-                new NonBodyParameter
+                new OpenApiParameter()
                 {
-                    In = "query",
+                    In = ParameterLocation.Query,
                     Name = _scheme.ParameterName,
                     Description = "The api version",
                     Required = true,
-                    Type = "string",
-                    Enum = new object[] {version}
+                    Schema = new OpenApiSchema
+                    {
+                        Type = "string",
+                        Enum = new List<IOpenApiAny>
+                        {
+                            new OpenApiString(version),
+                        },
+                    },
                 });
         }
     }
