@@ -4,22 +4,17 @@
 
 using System.Linq;
 using Microsoft.AspNetCore.Mvc.ApiExplorer;
-using Swashbuckle.AspNetCore.Swagger;
+using Microsoft.OpenApi.Models;
 using Swashbuckle.AspNetCore.SwaggerGen;
 
 namespace Microsoft.AspNetCore.ApiVersioning.Swashbuckle
 {
     public class SwaggerModelBindingOperationFilter : IOperationFilter
     {
-        public void Apply(Operation operation, OperationFilterContext context)
+        public void Apply(OpenApiOperation operation, OperationFilterContext context)
         {
-            foreach (IParameter parameter in operation.Parameters)
+            foreach (OpenApiParameter parameter in operation.Parameters)
             {
-                if (parameter.In.ToLower() == "modelbinding")
-                {
-                    parameter.In = "query";
-                }
-
                 // For some reason Swashbuckle doesn't honor ModelMetadata.IsRequired for parameters
                 ApiParameterDescription apiParamDesc =
                     context.ApiDescription.ParameterDescriptions.FirstOrDefault(p => p.Name == parameter.Name);
