@@ -4,9 +4,11 @@
 
 using System;
 using JetBrains.Annotations;
+using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.ApiExplorer;
 using Microsoft.AspNetCore.Mvc.ApplicationModels;
 using Microsoft.Extensions.DependencyInjection;
+using Microsoft.Extensions.Options;
 
 namespace Microsoft.AspNetCore.ApiVersioning
 {
@@ -18,9 +20,25 @@ namespace Microsoft.AspNetCore.ApiVersioning
             Action<ApiVersioningOptions> configureOptions)
         {
             return services.AddSingleton<VersionedControllerProvider>()
-                .AddTransient<IApplicationModelProvider, VersionedApiApplicationModelProvider>()
+                .AddSingleton<VersionedApiConvention>()
+                .AddSingleton<IConfigureOptions<MvcOptions>, ApiVersioningMvcOptionsSetup>()
                 .AddTransient<IApiDescriptionProvider, RequiredParameterDescriptorProvider>()
                 .Configure(configureOptions);
+        }
+    }
+
+    public class ApiVersioningMvcOptionsSetup : IConfigureOptions<MvcOptions>
+    {
+        private readonly VersionedApiConvention _apiConvention;
+
+        public ApiVersioningMvcOptionsSetup(VersionedApiConvention apiConvention)
+        {
+            _apiConvention = apiConvention;
+        }
+
+        public void Configure(MvcOptions options)
+        {
+            options.Conventions.Add(_apiConvention);
         }
     }
 }
