@@ -7,6 +7,7 @@ using JetBrains.Annotations;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore.Infrastructure;
 using Microsoft.EntityFrameworkCore.Metadata.Builders;
+using Microsoft.EntityFrameworkCore.Migrations;
 
 namespace Microsoft.DotNet.EntityFrameworkCore.Extensions
 {
@@ -34,21 +35,15 @@ namespace Microsoft.DotNet.EntityFrameworkCore.Extensions
             historyBuilder.Property<DateTime>("SysStartTime").HasColumnType("datetime2").ValueGeneratedOnAddOrUpdate();
             historyBuilder.Property<DateTime>("SysEndTime").HasColumnType("datetime2").ValueGeneratedOnAddOrUpdate();
 
-            historyBuilder.HasIndex("SysEndTime", "SysStartTime").ForSqlServerIsClustered();
+            historyBuilder.HasIndex("SysEndTime", "SysStartTime").IsClustered();
 
             return modelBuilder;
         }
 
-        public static IndexBuilder ForSqlServerIsColumnstore([NotNull] this IndexBuilder indexBuilder)
-        {
-            indexBuilder.Metadata.SetAnnotation(DotNetExtensionsAnnotationNames.Columnstore, true);
-            return indexBuilder;
-        }
-
         public static DbContextOptionsBuilder AddDotNetExtensions(this DbContextOptionsBuilder optionsBuilder)
         {
-            ((IDbContextOptionsBuilderInfrastructure) optionsBuilder).AddOrUpdateExtension(
-                new DotNetEntityFrameworkExtension());
+            optionsBuilder.ReplaceService<IMigrationsAnnotationProvider, SystemVersionedSqlServerMigrationsAnnotationProvider>();
+            optionsBuilder.ReplaceService<IMigrationsSqlGenerator, SystemVersionedSqlServerMigrationsSqlGenerator>();
             return optionsBuilder;
         }
     }
