@@ -1,7 +1,6 @@
 using System;
 using Microsoft.Build.Framework;
 using Microsoft.Extensions.Logging;
-using Microsoft.Extensions.Logging.Abstractions.Internal;
 
 namespace Microsoft.DotNet.Maestro.Tasks
 {
@@ -48,6 +47,23 @@ namespace Microsoft.DotNet.Maestro.Tasks
         public IDisposable BeginScope<TState>(TState state)
         {
             return NullScope.Instance;
+        }
+    }
+
+    /// <summary>
+    /// An empty scope without any logic
+    /// </summary>
+    internal class NullScope : IDisposable
+    {
+        public static NullScope Instance { get; } = new NullScope();
+
+        private NullScope()
+        {
+        }
+
+        /// <inheritdoc />
+        public void Dispose()
+        {
         }
     }
 }
