@@ -38,7 +38,7 @@ namespace SubscriptionActorService
         {
             services.AddSingleton<IActionRunner, ActionRunner>();
             services.AddSingleton<IMergePolicyEvaluator, MergePolicyEvaluator>();
-            services.AddSingleton<IRemoteFactory, DarcRemoteFactory>();
+            services.AddScoped<IRemoteFactory, DarcRemoteFactory>();
             services.AddSingleton<TemporaryFiles>();
             services.AddGitHubTokenProvider();
             services.AddAzureDevOpsTokenProvider();
