@@ -15,10 +15,7 @@ using Maestro.MergePolicies;
 using Microsoft.DotNet.DarcLib;
 using Microsoft.DotNet.ServiceFabric.ServiceHost;
 using Microsoft.DotNet.ServiceFabric.ServiceHost.Actors;
-using Microsoft.DotNet.Services.Utility;
-using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Logging;
-using Microsoft.Extensions.Logging.Abstractions.Internal;
 using Microsoft.ServiceFabric.Actors;
 using Microsoft.ServiceFabric.Actors.Runtime;
 using Microsoft.ServiceFabric.Data;
@@ -193,7 +190,7 @@ namespace SubscriptionActorService
             DarcRemoteFactory = darcFactory;
             ActionRunner = actionRunner;
             SubscriptionActorFactory = subscriptionActorFactory;
-            Logger = loggerFactory.CreateLogger(TypeNameHelper.GetTypeDisplayName(GetType()));
+            Logger = loggerFactory.CreateLogger(GetType());
         }
 
         public ILogger Logger { get; }
@@ -257,7 +254,7 @@ namespace SubscriptionActorService
         /// <returns>Build</returns>
         private Task<Build> GetBuildAsync(int buildId)
         {
-            return Context.Builds.FindAsync(buildId);
+            return Context.Builds.FindAsync(buildId).AsTask();
         }
 
         public Task RunProcessPendingUpdatesAsync()
