@@ -7,11 +7,20 @@ using System.Threading.Tasks;
 using Microsoft.DotNet.ServiceFabric.ServiceHost;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.DependencyInjection.Extensions;
+using Microsoft.Extensions.Logging;
+using Xunit.Abstractions;
 
 namespace SubscriptionActorService.Tests
 {
-    public class TestsWithServices : TestsWithMocks
+    public abstract class TestsWithServices : TestsWithMocks
     {
+        private readonly ITestOutputHelper _output;
+
+        protected TestsWithServices(ITestOutputHelper output)
+        {
+            _output = output;
+        }
+
         protected virtual void RegisterServices(IServiceCollection services)
         {
         }
@@ -26,6 +35,7 @@ namespace SubscriptionActorService.Tests
             var services = new ServiceCollection();
             Environment.SetEnvironmentVariable("ENVIRONMENT", "XUNIT");
             services.TryAddSingleton(typeof(IActorProxyFactory<>), typeof(ActorProxyFactory<>));
+            services.AddLogging(l => l.AddProvider(new XUnitLogger(_output)));
             RegisterServices(services);
             using (ServiceProvider container = services.BuildServiceProvider())
             using (IServiceScope scope = container.GetRequiredService<IServiceScopeFactory>().CreateScope())
