@@ -6,17 +6,18 @@ using Microsoft.DotNet.Darc.Options;
 using Microsoft.Extensions.Logging;
 using System;
 using System.Threading.Tasks;
+using Microsoft.Extensions.DependencyInjection;
+using Microsoft.Extensions.Logging.Console;
 
 namespace Microsoft.DotNet.Darc.Operations
 {
     internal abstract class Operation : IDisposable
     {
-        protected ILoggerFactory _loggerFactory;
-        private ILogger _logger;
+        private readonly ServiceProvider _provider;
 
-        protected ILogger Logger { get { return _logger; } }
+        protected ILogger<Operation> Logger { get; }
 
-        public Operation(CommandLineOptions options)
+        protected Operation(CommandLineOptions options)
         {
             // Because the internal logging in DarcLib tends to be chatty and non-useful,
             // we remap the --verbose switch onto 'info', --debug onto highest level, and the
@@ -30,32 +31,27 @@ namespace Microsoft.DotNet.Darc.Operations
             {
                 level = LogLevel.Information;
             }
-            _loggerFactory = new LoggerFactory().AddConsole(level);
-            _logger = _loggerFactory.CreateLogger<Operation>();
+
+            ServiceCollection collection = new ServiceCollection();
+            collection.AddLogging(b => b.AddConsole().AddFilter(l => l >= level));
+            _provider = collection.BuildServiceProvider();
+            Logger = _provider.GetRequiredService<ILogger<Operation>>();
         }
 
         public abstract Task<int> ExecuteAsync();
 
-        #region IDisposable Support
-        private bool disposedValue = false; // To detect redundant calls
-
         protected virtual void Dispose(bool disposing)
         {
-            if (!disposedValue)
+            if (disposing)
             {
-                if (disposing)
-                {
-                    _loggerFactory.Dispose();
-                }
-
-                disposedValue = true;
+                _provider?.Dispose();
             }
         }
-        
+
         public void Dispose()
         {
             Dispose(true);
+            GC.SuppressFinalize(this);
         }
-        #endregion
     }
 }
