@@ -13,7 +13,8 @@ namespace Microsoft.DotNet.Internal.DependencyInjection.Testing
         private static readonly ImmutableList<string> s_exemptTypes = ImmutableList.Create(
             "System.Fabric.ServiceContext",
             "Microsoft.Extensions.Options.IConfigureOptions`1",
-            "Microsoft.Extensions.Caching.Memory.MemoryCacheOptions"
+            "Microsoft.Extensions.Caching.Memory.MemoryCacheOptions",
+            "Microsoft.Extensions.DependencyInjection.IServiceScopeFactory"
         );
 
         private static readonly ImmutableList<string> s_exemptNamespaces = ImmutableList.Create(
@@ -45,7 +46,7 @@ namespace Microsoft.DotNet.Internal.DependencyInjection.Testing
                     continue;
                 }
 
-                if (!IsTypeResolvable(service.ImplementationType, services, allErrors))
+                if (!IsTypeResolvable(service.ImplementationType, services, allErrors, service.Lifetime))
                 {
                     allResolved = false;
                 }
@@ -57,7 +58,11 @@ namespace Microsoft.DotNet.Internal.DependencyInjection.Testing
             return allResolved;
         }
 
-        private static bool IsTypeResolvable(Type type, ServiceCollection services, StringBuilder msgBuilder)
+        private static bool IsTypeResolvable(
+            Type type,
+            ServiceCollection services,
+            StringBuilder msgBuilder,
+            ServiceLifetime serviceLifetime)
         {
             ConstructorInfo[] constructors = type
                 .GetConstructors(BindingFlags.Public | BindingFlags.Instance)
@@ -73,7 +78,7 @@ namespace Microsoft.DotNet.Internal.DependencyInjection.Testing
             string errorMessage = null;
             foreach (ConstructorInfo ctor in constructors)
             {
-                if (IsConstructorResolvable(ctor, services, errorMessage == null, out string ctorMsg))
+                if (IsConstructorResolvable(ctor, services, errorMessage == null, serviceLifetime, out string ctorMsg))
                 {
                     return true;
                 }
@@ -88,7 +93,12 @@ namespace Microsoft.DotNet.Internal.DependencyInjection.Testing
             return false;
         }
 
-        private static bool IsConstructorResolvable(ConstructorInfo ctor, ServiceCollection services, bool recordErrors, out string errorMessage)
+        private static bool IsConstructorResolvable(
+            ConstructorInfo ctor,
+            ServiceCollection services,
+            bool recordErrors,
+            ServiceLifetime serviceLifetime,
+            out string errorMessage)
         {
             errorMessage = null;
             bool resolvedAllParameters = true;
@@ -106,6 +116,22 @@ namespace Microsoft.DotNet.Internal.DependencyInjection.Testing
                 ServiceDescriptor parameterService = services.FirstOrDefault(s => IsMatchingServiceRegistration(s.ServiceType, p.ParameterType));
                 if (parameterService != null)
                 {
+                    if (serviceLifetime == ServiceLifetime.Singleton &&
+                        parameterService.Lifetime == ServiceLifetime.Scoped)
+                    {
+                        if (!resolvedAllParameters)
+                        {
+                            msgBuilder.Append(", ");
+                        }
+                        
+                        msgBuilder.Append("<SCOPED> ");
+                        msgBuilder.Append(p.Name);
+                        msgBuilder.Append(" of type ");
+                        msgBuilder.Append(GetDisplayName(p.ParameterType));
+
+                        resolvedAllParameters = false;
+                    }
+
                     continue;
                 }
 
@@ -140,7 +166,7 @@ namespace Microsoft.DotNet.Internal.DependencyInjection.Testing
                 // The name of IOptions<Pizza> is "IOptions`1"
                 // The full name has the other types, but they are all fully qualified (and also still have the `1 on them)
                 string baseName = type.Name.Split('`')[0];
-                return $"{baseName}<{string.Join(",", type.GetGenericArguments().Select(GetDisplayName))}>";
+                return $"{baseName}<{string.Join(", ", type.GetGenericArguments().Select(GetDisplayName))}>";
             }
 
             return type.Name;
